Task1

+++ Task_1_3.java
@@ -16,8 +16,8 @@ public class Task_1_3 {
         System.out.println("The sum of the numbers is " + (first + second + third));
         System.out.println("The product of the numbers is " + (first * second * third));
 
-        float avg = (float) (first + second + third) / 3;
-        System.out.println("The average of the numbers is " + String.format("%.6f", avg));
+        double avg = ((double) (first + second + third)) / 3.0;
+        System.out.println("The average of the numbers is " + String.format("%.13f", avg));
 
         scanner.close();
     }
